use ddtags instead of container_id
Tests also work now, although they use two different code paths depending on
weather you're in a container or not.

The only way I can think to get around this is to always run tests in a
container and remove the signal we use to determine if we're in a container.
I can do that, but only if it's worth the time/effort.